+++ inst/templates/usedrake/R/packages.R
@@ -1,0 +1,4 @@
+# Load all your packages before calling make().
+
+library(drake)
+library(tibble)